Move TimeSlot.java into Entities/Concrete directory
