+++ src/Server/Entities/Concrete/TimeSlot.java
@@ -1,4 +1,4 @@
-package Server.Entities;
+package Server.Entities.Concrete;
 
 import java.text.DateFormat;
 import java.text.ParseException;
